+++ src/FakeXrmEasy.Abstractions/CommercialLicense/StockKeepingUnits.cs
@@ -1,0 +1,49 @@
+namespace FakeXrmEasy.Abstractions.CommercialLicense
+{
+    /// <summary>
+    /// Known SKUs of products
+    /// </summary>
+    public enum StockKeepingUnits
+    {
+        /// <summary>
+        /// FXE Monthly Boutique Subscription
+        /// </summary>
+        FXE_BOU_1 = 1,
+        /// <summary>
+        /// FXE Monthly Team Subscription
+        /// </summary>
+        FXE_TEA_1 = 2,
+        /// <summary>
+        /// FXE Monthly Large Subscription
+        /// </summary>
+        FXE_LAR_1 = 3,
+        
+        /// <summary>
+        /// FXE Annual Boutique Subscription
+        /// </summary>
+        FXE_BOU_12 = 4,
+        /// <summary>
+        /// FXE Annual Team Subscription
+        /// </summary>
+        FXE_TEA_12 = 5,
+        /// <summary>
+        /// FXE Annual Large Subscription
+        /// </summary>
+        FXE_LAR_12 = 6,
+        
+        /// <summary>
+        /// FXE Business Plan
+        /// </summary>
+        FXE_BUS_12 = 7,
+        
+        /// <summary>
+        /// FXE Enterprise plan
+        /// </summary>
+        FXE_ENT_12 = 8,
+        
+        /// <summary>
+        /// FXE Ultimate plan
+        /// </summary>
+        FXE_ULT_12 = 9,
+    }
+}